Benself/apps/api/pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "benself-api"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "graphiti-core[kuzu]>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.4",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.30",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "pytest>=8.0.0"]

[project.scripts]
benself_api = "benself_api.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/benself_api"]

[tool.hatch.build.targets.sdist]
include = ["src/"]
